Move test to enc-string
